+++ ui/components-lib/src/components/HistoryView/DetailDrawer/_DetailDrawer.scss
@@ -1,4 +1,4 @@
-@use 'tokens' as *;
+@use '../tokens' as *;
 
 .hp-drawer {
   width: 0;
