+++ ui/components-lib/src/components/HistoryView/FlowCell/_FlowCell.scss
@@ -1,4 +1,4 @@
-@use 'tokens' as *;
+@use '../tokens' as *;
 
 .cell {
   appearance: none;
